Remove duplication in jest-playwright configuration
WIP shared config

clean config

simplify config

simplify config for e2e and perf tests

fix bundle tests config

improve log

remove todo
